input_chunk: use an array to store the route_mask value
This patch increases the size of the bitfield used to keep track of which
output plugins a chunk is routed to. The previous implementation would
misroute chunks if more than 32 output plugins were configured. This
implementation supports an arbitrary number of plugins, with the limit
defined by FLB_ROUTES_MASK_ELEMENTS

Signed-off-by: Joaquin Luna <jluna@tableau.com>