routes_mask: methods for maintaining a large bitfield in an array of values
Helper methods for getting, setting, and clearing bits in the router mask
bitfield. The bitfield is stored as an array of uint64_t values where each
bit is used to determine if an input chunk is to be routed to a given
output plugin. Each configured output plugin has a unique id that is assigned
sequentially. Each bit in the bitfield corresponds to the id of an output
plugin

Signed-off-by: Joaquin Luna <jluna@tableau.com>